ingestion/requirements.txt
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==47.0.0
db-dtypes==1.5.1
Faker==40.15.0
google-api-core==2.30.3
google-auth==2.49.2
google-auth-oauthlib==1.3.1
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
idna==3.13
numpy==2.4.4
oauthlib==3.3.1
packaging==26.2
pandas==2.3.3
pandas-gbq==0.35.0
proto-plus==1.27.2
protobuf==6.33.6
psutil==7.2.2
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydata-google-auth==1.9.1
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
requests-oauthlib==2.0.0
setuptools==82.0.1
six==1.17.0
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
